requirements.txt
fastapi
uvicorn
pymongo
dnspython
python-dotenv
passlib[bcrypt]
python-jose
email-validator
python-multipart

